% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acor.R
\name{acor}
\alias{acor}
\title{acor}
\usage{
acor(data, method = "r", dt = 30, stop = 0.5)
}
\arguments{
\item{data}{A timeseries data.frame with a timestamp column in first position}

\item{method}{Choose between calculation via "r" or "cpp"}

\item{dt}{Class-width in days}

\item{stop}{Stop calculation when reached a certain minimum}
}
\value{
A data.frame with class-wise levels of autocorrelation
}
\description{
Calculate the autocorrelation classwise for a timeseries data.frame
}
