% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acor_cpp.R
\name{acor_cpp}
\alias{acor_cpp}
\title{acor_r}
\usage{
acor_cpp(data, acf, stop)
}
\arguments{
\item{data}{a timeseries data.frame with a timestamp column in first position}

\item{acf}{A prepared data.frame to fill results in}

\item{stop}{Stop calculation when reached a certain minimum}
}
\value{
A data.frame with class-wise levels of autocorrelation
}
\description{
A function to actually calculate the autocorrelation classwise via c++
}
