% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acor_pairs}
\alias{acor_pairs}
\title{Creates pairs of data}
\arguments{
\item{data}{A data.frame. First column: date, second column: values.}

\item{FromTo}{Limits for pairing}
}
\value{
Two column matrix of paired data
}
\description{
Creates pairs of data
}
