% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acor_cpp.R, R/acor_r.R
\name{acor_r}
\alias{acor_r}
\title{acor_r}
\usage{
acor_r(data, dt = 30)

acor_r(data, dt = 30)
}
\arguments{
\item{data}{a timeseries data.frame with a timestamp column in first position}

\item{dt}{classwidth in days}
}
\value{
A data.frame with all the necessary information.

A data.frame with all the necessary information.
}
\description{
A function to actually calculate the autocorrelation classwise via R

A function to actually calculate the autocorrelation classwise via R
}
