% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acor_visualisation.R
\name{acor_visualisation}
\alias{acor_visualisation}
\title{acor_visualisation}
\usage{
acor_visualisation(accordf, step = 0.05)
}
\value{
A plot.
}
\description{
A function to show number of timeseries above certain acor level.
}
