\name{classwiseAcor-package}
\alias{classwiseAcor-package}
\alias{classwiseAcor}
\docType{package}
\title{\packageTitle{classwiseAcor}}
\description{\packageDescription{classwiseAcor}}
\section{Package Content}{\packageIndices{classwiseAcor}}
\author{\packageAuthor{classwiseAcor}}
\section{Maintainer}{\packageMaintainer{classwiseAcor}}
\keyword{package}
